Ficed the number of SH number in the log file. Changed the datatype of global_initial from consolidator.cpp/hpp to unsigned long long and changed the initial value to 0 to the total_count which was initlially initialized to global_initial

+++ ErrorFinder23.3/Consolidator.hpp
@@ -94,6 +94,6 @@ class Consolidator
 	std::string ibg_str;
 	std::string final_sh_str;
 	std::string ma_thresh_str;
-	int global_initial;
+	unsigned long long  global_initial;
 };
 #endif // ndef CONSOLIDATOR_H
